+++ lib/header/ControlPacketDisassembler.hpp
@@ -8,9 +8,9 @@ private:
 
     std::vector<IControlPacketDisassemblerObserver *> observers;
 
-    uint16_t packetDisassembly_getId(const std::vector<uint8_t> &packet);
+    uint32_t packetDisassembly_getId(const std::vector<uint8_t> &packet);
     uint32_t packetDisassembly_getCrc(const std::vector<uint8_t> &packet);
-    void packetDisassembly_processFlags(const std::vector<uint8_t> &packet, const uint8_t &id);
+    void packetDisassembly_processFlags(const std::vector<uint8_t> &packet, const uint32_t &id);
 
 public:
