+++ lib/header/InterthreadNotification.hpp
@@ -17,7 +17,7 @@ struct InterthreadNotification
     };
 
     const Type type;
-    const uint16_t referencedPacket_id;
+    const uint32_t referencedPacket_id;
 
-    InterthreadNotification(const Type &type, const uint16_t &referencedPacket_id = 0);
+    InterthreadNotification(const Type &type, const uint32_t &referencedPacket_id = 0);
 };
