+++ lib/implementation/InterthreadNotification.cpp
@@ -1,5 +1,5 @@
 #include "../header/InterthreadNotification.hpp"
 
-InterthreadNotification::InterthreadNotification(const Type &type, const uint16_t &referencedPacket_id)
+InterthreadNotification::InterthreadNotification(const Type &type, const uint32_t &referencedPacket_id)
     : type(type), referencedPacket_id(referencedPacket_id)
 {}
